[SPARK-44220][SQL] Move StringConcat to sql/api
### What changes were proposed in this pull request?

Move StringConcat to `sql/api` module.

### Why are the changes needed?

StringConcat is widely used in data types. As we plan to move entire data type family to sql/api, we should move StringConcat.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing UT

Closes #41764 from amaliujia/move_out_string_concat.

Authored-by: Rui Wang <rui.wang@databricks.com>
